dfs_audit_backup/requirements.txt
playwright>=1.40.0
mcp>=1.0.0
openai>=1.0.0
pyyaml>=6.0
aiofiles>=23.0.0
